+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
  - Add `ChunkGridTraits::chunks_in_array_subset()`
+ - Add `ArrayChunkCacheExt`, `ChunkCache`, `ChunkCacheLru{Size,Chunk}Limit`
 
 ### Changed
  - **Breaking**: `Arc` instead of `Box` partial decoders