+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
  - Add `ChunkGridTraits::chunks_in_array_subset()`
+ - Add `ArrayChunkCacheExt`, `ChunkCache`, `ChunkCacheLru{Size,Chunk}Limit`
 
 ### Changed
  - **Breaking**: `Arc` instead of `Box` partial decoders
@@ -16,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Specialise `set_partial_values` for `MemoryStore`
  - Bump maximum supported `ndarray` version from 0.15 to 0.16
  - **Breaking**: Make `create_chunk_grid_{regular,rectangular}` `pub(crate)` in alignment with other internal create from metadata methods
+ - **Breaking**: Bump MSRV to 1.76 (8 February, 2024)
 
 ### Fixed
  - `[async_]store_set_partial_values` no longer truncates